requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
easygui==0.98.3
et_xmlfile==2.0.0
h11==0.14.0
idna==3.10
numpy==2.2.0
odfpy==1.4.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PySocks==1.7.1
python-calamine==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pyxlsb==1.0.10
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Wand==0.6.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
